requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0.post1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
marshmallow==3.23.2
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.58.1
orjson==3.10.12
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pypdf==5.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
